experiment created
Created experiment, use command "make experiment" then "./experiment", and the program will automatically populate the csv files
