+++ Makefile
@@ -2,7 +2,10 @@ CC=gcc
 CFLAGS=-I.
 
 main: mt-collatz.cpp
-	$(CXX) $(CXXFLAGS) mt-collatz.cpp -o main
+	$(CXX) $(CXXFLAGS) mtCollatz.cpp -o main
+
+experiment: experiment.cpp
+	$(CXX) $(CXXFLAGS) experiment.cpp -o experiment
 
 run: $(TARGET)
 	./$(TARGET) $(ARGS)
