+++ data.csv
@@ -1,0 +1,80 @@
+20000000, 1, 11.752150
+20000000, 1, 1.000367
+20000000, 1, 1.000239
+20000000, 1, 1.000278
+20000000, 1, 1.000294
+20000000, 1, 1.000288
+20000000, 1, 1.000240
+20000000, 1, 1.000314
+20000000, 1, 1.000227
+20000000, 1, 1.000252
+20000000, 2, 1.000268
+20000000, 2, 1.000314
+20000000, 2, 1.000274
+20000000, 2, 1.000311
+20000000, 2, 1.000331
+20000000, 2, 1.000341
+20000000, 2, 1.000304
+20000000, 2, 1.000303
+20000000, 2, 1.000259
+20000000, 2, 1.000316
+20000000, 3, 1.000334
+20000000, 3, 1.000407
+20000000, 3, 1.000373
+20000000, 3, 1.000338
+20000000, 3, 1.000298
+20000000, 3, 1.000368
+20000000, 3, 1.000306
+20000000, 3, 1.000304
+20000000, 3, 1.000387
+20000000, 3, 1.000441
+20000000, 4, 1.000452
+20000000, 4, 1.000351
+20000000, 4, 1.000506
+20000000, 4, 1.000362
+20000000, 4, 1.000357
+20000000, 4, 1.000377
+20000000, 4, 1.000375
+20000000, 4, 1.000381
+20000000, 4, 1.000422
+20000000, 4, 1.000409
+20000000, 5, 1.000426
+20000000, 5, 1.000395
+20000000, 5, 1.000575
+20000000, 5, 1.000471
+20000000, 5, 1.000575
+20000000, 5, 1.000389
+20000000, 5, 1.000450
+20000000, 5, 1.000418
+20000000, 5, 1.000431
+20000000, 5, 1.000513
+20000000, 6, 1.000525
+20000000, 6, 1.000481
+20000000, 6, 1.000553
+20000000, 6, 1.000425
+20000000, 6, 1.000454
+20000000, 6, 1.000433
+20000000, 6, 1.000443
+20000000, 6, 1.000419
+20000000, 6, 1.000406
+20000000, 6, 1.000553
+20000000, 7, 1.000625
+20000000, 7, 1.000438
+20000000, 7, 1.000511
+20000000, 7, 1.000522
+20000000, 7, 1.000508
+20000000, 7, 1.000482
+20000000, 7, 1.000485
+20000000, 7, 1.000792
+20000000, 7, 1.000530
+20000000, 7, 1.000743
+20000000, 8, 1.000565
+20000000, 8, 1.000513
+20000000, 8, 1.000735
+20000000, 8, 1.000577
+20000000, 8, 1.000763
+20000000, 8, 1.000514
+20000000, 8, 1.000547
+20000000, 8, 1.000488
+20000000, 8, 1.000585
+20000000, 8, 1.000552
